♻️ Refactor: 자동 저장 기능 커스텀훅으로 분리
