+++ src/layouts/components/brainstormNote/BrainstormNote.jsx
@@ -1,3 +1,4 @@
+import { useEffect } from 'react';
 import IconButton from '../../../components/iconButton/IconButton';
 import { CancelIcon, PrevIcon } from '../../../components/iconButton/Icons';
 import { useNoteStore, VIEW_TYPE } from '../../../store/useNoteStore';
@@ -6,6 +7,13 @@ import NoteList from './NoteList';
 
 const BrainstormNote = ({ onClose }) => {
   const { view, setView } = useNoteStore();
+
+  // 언마운트뷰될때 리스트로 초기화
+  useEffect(() => {
+    return () => {
+      setView(VIEW_TYPE.LIST);
+    };
+  }, [setView]);
 
   const panelStyle = {
     boxShadow: '0px 0px 16px 0px rgba(217, 217, 217, 0.16)',
